Finish documenting runtime package
Some minor refactoring
