+++ lib/stringlib/packing.go
@@ -6,6 +6,7 @@ import (
 	"fmt"
 	"unsafe"
 
+	"github.com/arnodel/golua/luastrings"
 	rt "github.com/arnodel/golua/runtime"
 )
 
@@ -113,7 +114,7 @@ func unpack(t *rt.Thread, c *rt.GoCont) (rt.Cont, *rt.Error) {
 	if err == nil && c.NArgs() >= 3 {
 		n, err = c.IntArg(2)
 	}
-	i := rt.StringNormPos(pack, int(n)) - 1
+	i := luastrings.StringNormPos(pack, int(n)) - 1
 	if i < 0 || i > len(pack) {
 		err = rt.NewErrorS("#3 out of string")
 	}
