+++ runtime/callable.go
@@ -1,0 +1,6 @@
+package runtime
+
+// Callable is the interface for callable values.
+type Callable interface {
+	Continuation(*Thread, Cont) Cont
+}
